% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_oracle_ddim.R
\name{full_oracle_ddim}
\alias{full_oracle_ddim}
\title{Full oracle d-dimensional test for log-concavity}
\usage{
full_oracle_ddim(data, B, alpha, mu, p, compute_ts)
}
\arguments{
\item{data}{\eqn{n x d} data frame containing iid observations.
One row per observation.
We wish to test whether the underlying density is log-concave.}

\item{B}{Number of repeated subsamples for test statistic construction}

\item{alpha}{Significance level}

\item{mu}{Mean parameter from the underlying distribution
\eqn{(1-p) N(0, I_d) + p N(-mu, I_d)}}

\item{p}{Mixing parameter from the underlying distribution
\eqn{(1-p) N(0, I_d) + p N(-mu, I_d)}}

\item{compute_ts}{Indicator for whether to compute test statistic.
Set \code{compute_ts = 0} to stop early if rejection is guaranteed after some
b < B subsamples.
Set \code{compute_ts = 1} to perform all B subsamples and compute the
test statistic.}
}
\value{
List containing \code{test_stat} and \code{reject_null}.
\itemize{
\item \code{test_stat} --- If \code{compute_ts = 1}, this is the final test statistic,
averaged over B subsamples. If \code{compute_ts = 0}, this is NA.
\item \code{reject_null} --- Indicator that equals 1 if we reject H_0 at level
\code{alpha} and 0 if we do not reject H_0 at level \code{alpha}.
}
}
\description{
Run the full oracle d-dimensional test of
"H_0: true density is log-concave" versus
"H_1: true density is not log-concave."
This function assumes that the underlying distribution is a d-dimensional
mixture of two Normal distributions with the form
\eqn{(1-p) N(0, I_d) + p N(-mu, I_d)}. In addition, this approach assumes
that we know
the true underlying density. (Hence, this is a helpful theoretical
comparison, but it likely will not be used in practice.)
This method averages test statistics over B subsamples and
rejects H_0 if the average exceeds 1/alpha.
}
