% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_oracle_axis.R
\name{partial_oracle_axis}
\alias{partial_oracle_axis}
\title{Partial oracle, axis-aligned test for log-concavity}
\usage{
partial_oracle_axis(data, B, alpha)
}
\arguments{
\item{data}{\eqn{n x d} data frame containing iid observations.
One row per observation.
We wish to test whether the underlying density is log-concave.}

\item{B}{Number of repeated subsamples for test statistic construction}

\item{alpha}{Significance level}
}
\value{
List containing \code{reject_null}.
\itemize{
\item \code{reject_null} --- Indicator that equals 1 if we reject H_0 at level
\code{alpha} and 0 if we do not reject H_0 at level \code{alpha}.
}
}
\description{
Run the partial oracle axis-aligned test of
"H_0: true density is log-concave" versus
"H_1: true density is not log-concave."
This function computes the numerator density by fitting a mixture of
two d-dimensional Normal distributions. This is a "partial oracle"
because this approach uses the fact that the density is a mixture of
Normals, but it estimates all parameters of the mixture.
This method computes test statistics (averaged over B subsamples) on each of
the d axis directions. The test rejects H_0 if at least one of the test
statistics exceeds d/alpha.
}
