% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_oracle_randproj.R
\name{partial_oracle_randproj}
\alias{partial_oracle_randproj}
\title{Title}
\usage{
partial_oracle_randproj(data, B, n_proj, alpha, compute_ts)
}
\arguments{
\item{compute_ts}{}
}
\value{

}
\description{
Title
}
